Move TranslatedSelectBox
