+++ core/src/com/unciv/ui/popups/options/ModCheckTab.kt
@@ -15,7 +15,7 @@ import com.unciv.models.ruleset.validation.UniqueValidator
 import com.unciv.models.translations.tr
 import com.unciv.ui.components.widgets.ExpanderTab
 import com.unciv.ui.components.widgets.TabbedPager
-import com.unciv.ui.components.TranslatedSelectBox
+import com.unciv.ui.components.widgets.TranslatedSelectBox
 import com.unciv.ui.components.extensions.surroundWithCircle
 import com.unciv.ui.components.extensions.toLabel
 import com.unciv.ui.components.extensions.toTextButton
