+++ core/src/com/unciv/ui/screens/newgamescreen/MapParametersTable.kt
@@ -15,7 +15,7 @@ import com.unciv.logic.map.MapSize
 import com.unciv.logic.map.MapSizeNew
 import com.unciv.logic.map.MapType
 import com.unciv.ui.components.widgets.ExpanderTab
-import com.unciv.ui.components.TranslatedSelectBox
+import com.unciv.ui.components.widgets.TranslatedSelectBox
 import com.unciv.ui.components.UncivSlider
 import com.unciv.ui.components.UncivTextField
 import com.unciv.ui.components.WrappableLabel
